Fix flogging a dead horse when sessions are unresponsive
As c.session was never set to nil, if the GCAC failed and a new
session failed to establish, we left the known-bad session in
place for the next scrape to needlessly test again. This led to
a negative number of open sessions as each one was closed multiple
times.